app/src/main/java/com/example/obes/model/Book/Book.java
package com.example.obes.model.Book;

public class Book {
    private int id;
    private String title;
    private String description;
    private String category;
    private boolean available;
    private int coverResourceId;
    private String author;
    private double price;
    private String condition;

    public Book(int id, String title, String description, String category, boolean available, int coverResourceId, String author, double price, String condition) {
        this.id = id;
        this.title = title;
        this.description = description;
        this.category = category;
        this.available = available;
        this.coverResourceId = coverResourceId;
        this.author = author;
        this.price = price;
        this.condition = condition;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }

    public boolean isAvailable() {
        return available;
    }

    public void setAvailable(boolean available) {
        this.available = available;
    }
    public boolean getAvailable() {
        return this.available;
    }

    public int getCoverResourceId() {
        return coverResourceId;
    }

    public void setCoverResourceId(int coverResourceId) {
        this.coverResourceId = coverResourceId;
    }

    public String getAuthor() {
        return author;
    }

    public void setAuthor(String author) {
        this.author = author;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public String getCondition() {
        return this.condition;
    }

    public void setCondition(String condition) {
        this.condition = condition;
    }
}
